Move EditPresetScreen to separate module
